[core] add placeholder files, rename parseCourse - go will run errors with empty files - consistent naming scheme with renaming parseCourse to readlesson - add test data files, next commit will be test
