+++ biocterm/setup/dev-setup.sh
@@ -1,0 +1,6 @@
+#!/bin/sh/
+
+// various golang repos and packages used are listed here 
+go get github.com/gravitational/configure 
+go get github.com/tools/godep
+go get gopkg.in/yaml.v2
